+++ cmake/custom/api.cmake
@@ -1,2 +1,2 @@
 add_subdirectory(api)
-include_directories(${PROJECT_SOURCE_DIR}/api)
+include_directories(${PROJECT_SOURCE_DIR}/numgrid)
